+++ spec/kitchen_spec/directions/bake_footnotes/main_spec.rb
@@ -4,7 +4,9 @@ require 'spec_helper'
 
 RSpec.describe Kitchen::Directions::BakeFootnotes do
   it 'calls v1' do
-    expect_any_instance_of(Kitchen::Directions::BakeFootnotes::V1).to receive(:bake).with(book: 'blah', number_format: :arabic)
+    expect_any_instance_of(Kitchen::Directions::BakeFootnotes::V1).to receive(:bake).with(
+      book: 'blah', number_format: :arabic, selector: nil
+    )
     described_class.v1(book: 'blah')
   end
 end